+++ src/main.rs
@@ -83,6 +83,7 @@ fn main() -> Result<(), String> {
                  --exclude -e [PACKAGE]... 'Package id specifications to exclude from coverage. See cargo help pkgid for more info'
                  --exclude-files [FILE]... 'Exclude given files from coverage results has * wildcard'
                  --timeout -t [SECONDS] 'Integer for the maximum time in seconds without response from test before timeout (default is 1 minute).'
+                 --follow-exec 'Follow executed processes capturing coverage information if they're part of your project.'
                  --release   'Build in release mode.'
                  --no-run 'Compile tests but don't run coverage'
                  --locked 'Do not update Cargo.lock'