+++ CHANGELOG.md
@@ -6,6 +6,7 @@ file.
 ## [Unreleased]
 ### Added
 - Added `--color` option matching cargo arg
+- `--follow-exec` option making exec tracing non-default
 
 ### Changed
 - Check through memory map for the first entry belonging to the executable [FIX]
@@ -13,7 +14,7 @@ file.
 
 ### Removed
 
-## [0.17.0] - 2020-11-10
+## [0.17.0] - 2020-11-10 [YANKED]
 ### Added
 - Now trace into executed binaries
 - Added `--avoid-cfg-tarpaulin` flag to remove `--cfg=tarpaulin` from the